set up a dedicated file for socket events
